# feat: Speed up `get_color_contrast()`