Optimizes `get_color_contrast()` by avoiding re-using `bs_get_contrast()` and instead directly creating a small Sass layer with a minimal set of Sass files.

This ends up speeding up `get_color_contrast()` calls by about 25x:

``` r
pkgload::load_all()
#> ℹ Loading bslib

results <- bench::mark(
  before = get_color_contrast_old("blue"),
  after = get_color_contrast("blue"), 
  min_time = 2
)

plot(results)
```

![image](https://github.com/user-attachments/assets/d3d81589-2d07-4e68-a776-ae7d81a74b4f)


``` r
results
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 before         29ms  30.57ms      32.3    6.22MB     51.1
#> 2 after         979µs   1.11ms     818.    98.81KB     23.1
```